Main

Description

Um módulo contendo definições Haskell para a compressão e descompressão de um tabuleiro de jogo de Bomberman.

Documentation

main :: IO ()

Função main fornecida no enunciado, que irá comprimir ou descomprimir a uma lista de strings ou uma string, respetivamente, de acordo com os inputs fornecidos.

encode :: [String] -> String

A função encode irá pegar numa lista de strings que corresponde ao tabuleiro de jogo e executará a função compress, para reduzir a lista para o menor número de caracters possíveis.

compress :: [String] -> [String]

A função compress é utilizada pela função encode para executar os processos de alteração ao tabuleiro necessários antes de se executar o unlines dele.

retiraP :: [String] -> [String]

A função retiraP é chamada na compress para retirar as paredes laterais do tabuleiro. Para tal, utiliza as funções retiraV e retiraH. Strings que descrevam as locações de power ups, jogadores ou bombas permanecerão inalteradas.

retiraV :: [String] -> [String]

A função retiraV é utilizada na retiraP e o seu objetivo é retirar os primeiros e últimos elementos de todas as strings das listas que descrevam o centro do tabuleiro.

retiraH :: [String] -> [String]

A função retiraH é utilizada na retiraP e o seu objetivo é retirar a primeira e última linha do tabuleiro.

decode :: String -> [String]

A função decode fará o processo contrário ao da função encode, ou seja, pega numa string comprimida pela encode e reverte o processo de compresso, restaurando o tabuleiro de jogo ao seu estado inicial.

decompress :: [String] -> [String]

A função decompress inclui todos os processos de descompressão a que o tabuleiro é sujeito após o lines dele na função decode.

adicionaP :: [String] -> [String]

A função adicionaP é chamada na função decompress e serve para adicionar as paredes laterais anteriormente retiradas pela retiraP.

adicionaV :: [String] -> [String]

A adicionaV é invocada pela adicionaP para restaurar as paredes verticais ao centro do tabuleiro.

adicionaH :: [String] -> Int -> [String]

A adicionaH é invocada pela adicionaP para restaurar a primeira e última linha ao tabuleiro.

linhaPU :: Int -> String

A função linhaPU, tal como na Tarefa 1, é utilizada para produzir uma linha preenchida unicamente por pedras, que será igual à primeira e última linha do tabuleiro original.