Main

Description

Um módulo contendo definições Haskell para a passagem de tempo num tabuleiro de jogo de Bomberman.

Documentation

main :: IO ()

Função main fornecida no enunciado, que recebe como inputs uma descrição de um mapa e o número de instantes de tempo restantes e usará esses dados para avançar o tempo do jogo e fechar o mapa, caso necessário.

avanca :: [String] -> Int -> [String]

A função avanca recebe a informação do mapa de jogo e o tempo restante para o jogo acabar, chamando depois várias funções que irão avancar o tempo do jogo, e dependendo do tempo restante, fechar o mapa.

removeB :: [String] -> [String]

A função removeB recebe a descrição de um mapa e remove, nessa descrição, todas as bombas cujos timers já tenham atingido 0.

aTempo :: [String] -> [String] -> [String]

A aTempo é chamada pela avanca para reduzir o tempo de todas as bombas no mapa. Para tal, utiliza uma função auxiliar.

auxAT :: [String] -> (Int, Int, Int) -> Int -> [String]

A função auxAT é a função auxiliar chamada pela aTempo e é a função que controla as explosões. Recebe o mapa e as coordenadas da bomba, assim como o raio da bomba e chama as funções subTij e subJog para mudarem a informação do mapa e dos jogadorespower upsbombas, respetivamente, juntando depois o resultado das duas funções.

subJog :: [String] -> (Int, Int, Int) -> Int -> [String]

A subJog é chamada pela auxAT para remover os tijolos que são atigidos quando ocorre uma explosão. Para tal, utiliza uma função auxiliar - auxSJ.

auxSJ :: ([String], [String]) -> [String] -> (Int, Int, Int) -> [String]

A auxSJ é chamada pela subJog, recebendo a informação do jogo, dividida em duas partes: a primeira parte corresponde ao próprio tabuleiro e a segunda parte à informação de jogadores, power ups e bombas. A auxSJ chama duas funções, a preJogH e a subJogV, para tratar da remoção dos jogadores, bombas e power ups, horizontalmente e verticalmente.

preJogH :: String -> [String] -> Int -> Int -> Int -> Int -> [String]

A preJogH é a função chamada pela auxSJ para preparar a informação que será fornecida à subJogH. Para tal, divide a string que lhe é fornecida em duas partes e chama uma função auxiliar.

auxPJogH :: (String, String) -> [String] -> Int -> Int -> Int -> Int -> [String]

A auxPJogH é a função auxiliar chamada pela preJogH, que recebendo as duas partes da string que será analisada, chama a subJogH para as duas metades assim como a locJog para o carácter do centro.

subJogH :: String -> [String] -> Int -> Int -> Int -> Int -> Int -> [String]

A função subJogH é utilizada para percorrer uma lista de carácteres, verificando se existem power ups, jogadores ou bombas no caminho da explosão, removendo-as caso existirem e parando a função caso encontre uma pedra ou tijolo.

subJogV :: [String] -> [String] -> Int -> Int -> Int -> Int -> Int -> [String]

A subJogV tem a função de percorrer uma parte do tabuleiro de jogo, verificando se existem jogadores, power ups ou bombas numa posição específica ao longo do tabuleiro.

locJog :: [String] -> (Int, Int, Int) -> [String]

A função locJog é utilizada em várias funções para verificar se numa coordenada específica existem jogadores, bombas ou power ups. No caso dos jogadores e power ups, estes serão removidos da informação do mapa, mas no caso das bombas o seu temporizador será reduzido para 1 caso este seja maior que 1. Caso a locJog seja chamada com c == 1, se a linha que estiver a ser analizada corresponda a uma bomba, esta será removida em vez de ter o seu temporizador reduzido como normalmente.

reduzB1 :: String -> Int -> String

A reduzB1 é chamada pela locJog para reduzir o timer de uma bomba para 1.

subTij :: [String] -> (Int, Int, Int) -> Int -> [String]

A subTij, semelhantemente à subJog é utilizada para atualizar a informação de jogo quando ocorre uma explosão, mas em vez de remover informação correspondente a jogadores, bombas ou power ups, tenta apenas remover os primeiros tijolos atingidos pelas tais explosões. Para este efeito, utiliza a mesma tática da subJog, de dividir o mapa e analizá-lo por partes.

auxST :: ([String], [String]) -> Int -> Int -> [String]

A função auxST é chamada pela subTij, recebendo o mapa dividido em partes, e chamando as funções subTijV e subTijH com estas partes.

subTijV :: [String] -> Int -> Int -> Int -> [String]

A subTijV é utilizada para analizar o mapa verticalmente. Passa pelas strings que lhe são fornecidas e vê se numa posição específica (a coordenada horizontal onde a bomba explodiu), se existe um tijolo. Caso exista, será removido.

auxSTV :: String -> Int -> Int -> String

A auxSTV é utilizada como auxiliar na subTijV para mudar as linhas onde se encontram tijolos.

subTijH :: String -> Int -> Int -> String

A subTijH é chamada pela auxST para analizar o mapa horizontalmente. Para tal, recebe uma linha do mapa (a linha onde a bomba explodiu) e analiza a linha, tentando encontrar nela tijolos.

auxSTH :: (String, String) -> Int -> Int -> String

A auxSTH é chamada pela subTijH, recebendo uma linha do tabuleiro dividida em duas partes, uma correspondendo à parte à esquerda da explosão e outra parte incluíndo a direita da explosão e o centro. Com estas partes, chama a subTHLR, onde irão ser subsitituídos os tijolos, caso necessário.

subTHLR :: String -> Int -> Int -> String

A subTHLR é usada na auxSTH e é usada para percorrer as partes do mapa que lhe são atribuidas a ver se encontra tijolos, que serão substituídos por espaços.

rCRaio :: String -> (String, String, String) -> Int -> (Int, Int, Int)

A função rCRaio é chamada pela função aTempo para retirar as coordenadas e raio de uma linha de informação correspondente a uma bomba cujo timer seja 0.

rCoords :: String -> (String, String) -> Int -> (Int, Int)

A rCoords é usada pela locJog para retirar as coordenadas de uma linha que descreva uma bomba, posição de jogador ou power up.

rTimer :: String -> String -> Int -> Int

A função rTimer é usada para retirar o valor do timer de uma linha de descrição de bomba.

converte :: String -> Int -> Int

A função converte é chamada em funções que retiram dados de linhas de informação do mapa, e é utilizada para converter os dados obtidos de Strings para Integers.

passaTempo :: [String] -> [String]

A passaTempo é a função que passa pelo mapa para verificar quais linhas têm bombas, aplicando a essas linhas a função reduzBomba antes de as devolver.

reduzBomba :: String -> Int -> String

A função reduzBombas é chamada pela passaTempo, recebendo as strings que a passaTempo encontra para que possa então reduzir o valor do timer dentro delas por um valor.

preSpiral :: [String] -> Int -> Int -> [String]

A preSpiral é a função responsável por preparar os inputs necessários para iniciar o processo de fechar o mapa. Para tal, chama a função spiral (para colocar as pedras) e a montaS (para voltar a juntar os dados que a preSpiral separa, assim como para chamar a locJog).

montaS :: ([String], (Int, Int)) -> [String] -> Int -> [String]

A montaS recebe o tabuleiro de jogo alterado pela spiral assim como as coordenadas onde a última pedra foi colocada. Chama a função recS para restaurar o tabuleiro alterado pela spiral e a locJog para retirar informações de jogadores, bombas ou power ups na posição onde foi colocada uma pedra e depois junta os outputs produzidos por essas duas funções.

recS :: [String] -> Int -> [String]

A recS é usada para reconstruir um tabuleiro de jogo após a função spiral reduzir a sua dimensão, adicionando linhas so com um carácter ao início e final até o comprimento ser igual ao pretendido. Após isso, irá chamar a função addRS para completar o trabalho.

addRS :: [String] -> Int -> [String]

A addRS é chamada pela recS para encher as linhas adicionadas pela recS até que todas as linhas tenham o comprimento desejado. Para tal, utiliza uma função auxiliar.

auxRS :: String -> Int -> String

A auxRS é usada como auxiliar na função addRS para adicionar os carácteres em falta às linhas que será necessário alterar na função addRS.

spiral :: [String] -> Int -> Int -> Int -> Int -> Int -> Int -> ([String], (Int, Int))

A função spiral é a função responsável por colocar uma pedra numa coordenada do mapa enquanto este fecha. Para tal, corre uma série de verificações para ver se as linhas ou colunas que está a analizar estão cheias - caso a parte exterior do mapa esteja toda preenchida por pedras, esta será removida e será analisado o interior do mapa, sendo a parte removida depois restaurada pela função recS. Caso a função spiral encontre uma linha ou coluna que não esteja preenchida, irá utilizar as funções spiralH ou spiralV para colocar uma pedra.

spiralH :: String -> Int -> Int -> String

A spiralH recebe a linha que se pretende atualizar e adiciona uma pedra nas coordenadas pedidas, excepto no caso onde nessa coordenada já existe uma pedra, onde não será realizada alteração.

spiralV :: [String] -> Int -> Int -> Int -> [String]

A spiralV recebe o tabuleiro inteiro utilizado na spiral e tenta adicionar uma pedra numa coluna específica, caso não exista já nessas coordenadas uma pedra.

contaSH :: String -> Int -> Int

A contaSH é utilizada para contar o número de pedras numa linha do mapa.

contaSV :: [String] -> Int -> Int -> Int

A contaSV é utilizada para contar o número de pedras numa coluna do mapa.

auxSV :: String -> Int -> String

A auxSV é uma função auxiliar da spiralV utilizada para colocar uma pedra no sítio indicado pelo input.

checkLinhaH :: String -> Bool

A checkLinhaH é utilizada para verificar se uma linha do mapa é formada por pedras exclusivamente.

checkLinhaV :: [String] -> Int -> Bool

A checkLinhaV é utilizada para verifcar se uma coluna do mapa é formada por pedras exclusivamente.

retiraP :: [String] -> [String]

A função retiraP (retirada da Tarefa 3) é chamada na compress para retirar as paredes laterais do tabuleiro. Para tal, utiliza as funções retiraV e retiraH. Strings que descrevam as locações de power ups, jogadores ou bombas permanecerão inalteradas.

retiraV :: [String] -> [String]

A função retiraV (Tarefa 3) é utilizada na retiraP e o seu objetivo é retirar os primeiros e últimos elementos de todas as strings das listas que descrevam o centro do tabuleiro.

retiraH :: [String] -> [String]

A função retiraH (Tarefa 3) é utilizada na retiraP e o seu objetivo é retirar a primeira e última linha do tabuleiro.