Main

Description

Um módulo contendo definições Haskell para a execução de um comando num jogo de Bomberman.

Documentation

move :: [String] -> Int -> Char -> [String]

A função move, após verificar que os comandos fornecidos são válidos, executa a função fazJogada, que irá fazer as modificações necessários ao tabuleiro de jogo.

main :: IO ()

Função main fornecida no enunciado, que chamará a função move para realizar uma acção num jogo de Bomberman.

fazJogada :: [String] -> Int -> Char -> [String]

A função fazJogada é chamada pela move, com o tabuleiro de jogo, o número do jogador a realizar uma acção e o tipo de acção que o jogador pretende realizar. A função utiliza o verificaJogador para determinar se a jogada se pode realizar, após a qual irá usar a checkRep e a insereJogada para produzir o output necessário.

verificaJogador :: [String] -> Int -> Bool

A verificaJogador é usada para verificar se o jogador indicado na chamada da função main/move existe no tabuleiro jogo. Nesse caso, devolverá True; se o jogador não estiver presente devolverá False.

checkRep :: [String] -> [String]

A função checkRep é chamada pela fazJogada para verificar o mapa gerado pelo insereJogada para verificar se numa coordenada existem várias coisas que não podem coexistir. Por exemplo, se um jogador se encontrar nas coordenadas (3,3) e um power up Bombas também existir nessas coordenadas, a checkRep removeria a linha + 3 3, que é correspondente ao power up.

repPU :: (Char, Char) -> [String] -> Bool

A repPU é chamada como auxiliar na checkRep para remover power ups presentes nas mesmas coordenadas que jogadores.

repB :: (Char, Char) -> [String] -> Bool

A repB é usada como auxiliar na checkRep para remover bombas presentes nas mesmas coordenadas.

insereJogada :: [String] -> Int -> String -> [String]

A função insereJogada recebe a jogada produzida pela função gereJogada e decide onde será inserida a jogado dependendo do tipo de jogada que é.

insereBomba :: [String] -> String -> [String]

A insereBomba é chamada pela insereJogada caso a jogada recebida começe com * (logo, seja uma bomba) e decide em que posição do tabuleiro será posta a jogada (depois do mapa, depois dos power ups, mas antes dos jogadores) decidindo também em que posição será posta comparativamente com as outras bombas.

compBomba :: (Char, Char) -> (Char, Char) -> Bool

A compBomba é chamada pela insereBomba para comparar duas descrições do tipo de bomba: uma que será inserida no tabuleiro e outra já la presente, devolvendo True caso a que se pretende inserir estiver numa posição anterior à já posta e False caso contrário.

insereUDLR :: [String] -> Int -> String -> [String]

A insereUDLR é a alternativa à insereBomba, ou seja é realizada caso a jogada realizada não seja a colocação de uma bomba. Esta função coloca a linha da jogada no sítio do mapa onde esteve a ultima jogada desse mesmo jogador.

gereJogada :: [String] -> Int -> Char -> (Int, Int, String) -> String

A gereJogada é invocada pela função fazJogada para produzir a jogada. Para tal utiliza cinco funções para os diferentes tipos de jogadas. moveUp para U, moveDown para D, moveLeft para L, moveRight para R e moveBomba para B.

moveUp :: [String] -> Int -> (Int, Int, String) -> Int -> Int -> String

A moveUp é invocada pela gereJogada caso o tipo de jogada fornecida seja U. Irá mudar as coordenadas do jogador, caso possível, para que se encontre na linha em cima.

moveDown :: [String] -> Int -> (Int, Int, String) -> Int -> Int -> String

A moveDown é invocada pela gereJogada caso o tipo de jogada fornecida seja D. Irá mudar as coordenadas do jogador, caso possível, para que se encontre na linha em baixo.

moveLeft :: [String] -> Int -> (Int, Int, String) -> Int -> Int -> String

A moveLeft é invocada pela gereJogada caso o tipo de jogada fornecida seja L. Irá mudar as coordenadas do jogador, caso possível, para que se encontre na mesma linha, na posição anterior.

moveRight :: [String] -> Int -> (Int, Int, String) -> Int -> Int -> String

A moveRight é invocada pela gereJogada caso o tipo de jogada fornecida seja R. Irá mudar as coordenadas do jogador, caso possível, para que se encontre na linha em cima, na posição seguinte

atualiza :: Int -> (Int, Int, String) -> Char -> String

A função atualiza é invocada pelas funções move direcionais e utiliza os dados gerados por essas funções para criar a linha que será adicionada ao mapa.

moveBomba :: [String] -> Int -> (Int, Int) -> String

A função moveBomba é chamada pela gereJogada para gerar uma linha de jogo que contenhana a descrição de uma bomba. Para verificar se a jogada é válida recorre à checkBomba.

checkBomba :: [String] -> Int -> Bool

A checkBomba é utilizada na moveBomba e serve para verificar se o jogador que pretende colocar a bomba já colocou outra bomba no mapa. Caso tenha, devolverá True.

addLB :: (Int, Int) -> Int -> Int -> Int -> String

A função addLB é chamada pela moveBomba caso não exista já uma bomba colocada no mapa pelo mesmo jogador e criará a descrição da bomba que se pretende juntar ao mapa. Junta as coordenadas, o número do jogador, o raio e o tempo numa string.

checkPU :: [String] -> (Int, Int) -> Char

A checkPU é utilizada pelas funções move direcionais para verificar qual power up é que o jogador está a apanhar.

rDados :: Char -> String -> (Int, Int) -> Int -> Char

A rDados é utilizada pela checkPU para retirar o power up que existe nas coordenadas onde o jogador se encontra.

obtemVals :: [String] -> Int -> Int -> Int -> String -> (Int, Int, String)

A obtemVals é utilizada pela fazJogada para retirar as coordenadas necessárias da descrição do jogador que se encontra a jogar, assim como a lista de power ups que esse jogador já apanhou em fases anteriores do jogo.

obtemDX :: String -> Int -> Int

A função obtemDX é chamada na obtemVals para retirar a coordenada correspondente à posição horizontal do jogador que se encontra a jogar.

obtemDY :: String -> Int -> Int

A função obtemDY é chamada na obtemVals para retirar a coordenada correspondente à posição vertical do jogador que se encontra a jogar.

obtemDS :: String -> String -> Int -> String

A função obtemDS é chamada na obtemVals para retirar os power ups que o jogador que se encontra a jogar já apanhou em rondas anteriores.