Main

Description

Um módulo contendo definições Haskell para a geração de um mapa de Bomberman.

Documentation

mapa :: Int -> Int -> [String]

A função mapa, após verificar se os dados fornecidos são validos, chama as funções addTijolos, addBombas e addFlames para criar o mapa.

mapaAux :: Int -> [String] -> [Int] -> [String]

A função mapaAux é chamada pela função mapa, e é utilizada para simplificar o trabalho feito pela mapa. Guarda os valores de addTijolos, addBombas e addFlames em variáveis para reduzir o espaço ocupado pelo código. Esta função gera a versão final do mapa, que será dado como output.

main :: IO ()

Função main fornecida no enunciado, que chamará a função mapa para poder produzir o mapa de Bomberman.

addTijolos :: [String] -> [Int] -> Int -> Int -> [String]

A addTijolos é chamada pela mapa e, com o mapa gerado pela função mapaVazio (função que gera uma versão do mapa contendo apenas pedras e espaços) e com a lista de números gerados pela powerUps (função que gera a lista de números que serão distribuídos pelo mapa como tijolos/power ups) irá substituir os espaços no mapa por tijolos, onde adequado.

auxT1 :: String -> [Int] -> Int -> Int -> String

A auxT1 é chamada pela addTijolos e serve para substituir os espaços por tijolos nas linhas do mapa onde se tem que preservar os cantos, ou seja, a primeira, segunda, penúltima e última linhas do mapa (não contando as paredes horizontais).

auxT2 :: String -> [Int] -> String

A auxT2, tal como a auxT1, é chamada pela addTijolos para substituir espaços, excepto que a auxT2 é encarregada de substituir espaços em linhas onde não hajam cantos para preservar.

numEsp :: String -> Int -> Int

A numEsp é chamada pela addTijolos, addBombas e addFlames, e conta o número de espaços ou tijolos contidos numa linha do mapa que não contenha cantos. 3

addBombas :: [String] -> [Int] -> Int -> Int -> [String]

A addBombas é chamada na mapaAux e é usada para criar uma lista de strings das posições de power ups bombas no mapa modificado gerado pela função addTijolos.

auxB1 :: String -> [Int] -> Int -> Int -> Int -> [String]

A auxB1 é uma função auxiliar utilizada na addBombas para criar linhas com a descrição da existência de bombas na primeira, segunda, penúltima e última linhas do mapa (não contanto com as paredes horizontais), tendo então em consideração os cantos do mapa, em que não poderão haver power ups.

auxB2 :: String -> [Int] -> Int -> Int -> Int -> [String]

A auxB2 é identica à função auxB1, sendo também chamada pela addBombas, com a excepção que não considera os contas que é necessário preservar no mapa, ou seja, é utilizada para descrever bombas presentes no mapa entre a terceira e antepenúltimas linhas do mapa (não contando com as paredes horizontais), inclusive.

addFlames :: [String] -> [Int] -> Int -> Int -> [String]

A função addFlames é chamada pela addTijolos, e tal como a addBombas, é utilizada para adicionar linhas ao final do mapa, nas quais são descritas as posições de todos os power ups da categoria Flames. Para tal, utiliza as mesmas técnicas que a função addBombas, incluindo a utilizão de duas funções auxiliares.

auxF1 :: String -> [Int] -> Int -> Int -> Int -> [String]

A auxF1 é a primeira auxiliar chamada pela addFlames, na qual são criadas as descrições de localizações de power ups da categoria Flames na primeira, segunda, penúltima e última linhas do mapa (não contando com as paredes horizontais).

auxF2 :: String -> [Int] -> Int -> Int -> Int -> [String]

A auxF2 é a segunda auxiliar utilizada na addFlames, na qual são geradas as posições dos power ups Flames nas linhas do mapa entre a terceira e antepenúltima linhas do mapa (não contando com as paredes horizontais), inclusive.

mapaVazio :: Int -> Int -> [String]

A mapaVazio é uma função chamada pela addTijolos, na qual é gerado um mapa de Bomberman vazio, ou seja, contem exclusivamente espaços e pedras, não contendo então power ups e/ou tijolos. A mapaVazio utiliza três funções auxiliares para gerar os três tipos de linhas presentes no mapa.

linhaPU :: Int -> String

A função linhaPU é utilizada na mapaVazio para gerar a primeira e última linha do mapa.

linhaMP :: Int -> Int -> String

A linhaMP é chamada pela mapaVazio para gerar as linhas do mapa em que não existem pedras a meio.

linhaMI :: Int -> Int -> String

A linhaMI é chamada pela mapaVazio para gerar as linhas do mapa em que existem pedras a meio.

powerUps :: Int -> Int -> [Int]

A função powerUps é chamada pela mapa e pela addTijolos para gerar a lista de números que, quando comparados com o mapa criado pela addTijolos, irão decidir onde estarão os tijolos, e em quais desses tijolos estarão escondidos power ups.

numPowerUps :: Int -> Int

A numPowerUps é utilizada pela powerUps para decidir quantos números a função powerUps precisará de gerar.